fix(#350): extract ASK_EXAMPLE_PROMPTS to satisfy react-refresh + use proper ul/li semantics
Two reviewer findings on PR #365:

1. CRITICAL — `react-refresh/only-export-components` (CI lint, --max-warnings=0)
   The non-primitive `ASK_EXAMPLE_PROMPTS: readonly string[]` was exported
   alongside React components in `AskMode.tsx`. The repo's
   `allowConstantExport: true` only whitelists string/number/boolean/template-
   literal exports — arrays don't qualify. Moved the constant to its own file
   `ask-example-prompts.ts` and updated all consumers (`AskMode.tsx`,
   `modes/index.ts`, `AskMode.test.tsx`).

2. WARNING — ARIA misuse in `AskExamplePrompts`
   The example-prompt `<button>`s had `role="listitem"`, which strips the
   implicit "button" role and makes screen readers announce "listitem"
   instead of "button". Replaced the `<div role="list">` + `<button role="listitem">`
   pattern with semantic `<ul>` / `<li>` wrapping the buttons. The buttons now
   keep their native button role and the list structure is conveyed through
   real DOM elements.

All 10 AskMode unit tests still pass; `data-testid="ask-example-prompt"` and
the `ASK_EXAMPLE_PROMPTS` shape are unchanged so RTL queries continue to work.
